created PJsonFormatter, restructured code to fix cyclic dependency issue
