+++ Src/PCompiler/PCommandLine/Options/PCheckerOptions.cs
@@ -70,7 +70,6 @@ namespace Plang.Options
             advancedGroup.AddArgument("graph-bug", null, "Output a DGML graph of the iteration that found a bug", typeof(bool));
             advancedGroup.AddArgument("graph", null, "Output a DGML graph of all test iterations whether a bug was found or not", typeof(bool));
             advancedGroup.AddArgument("xml-trace", null, "Specify a filename for XML runtime log output to be written to", typeof(bool));
-            advancedGroup.AddArgument("json-trace", null, "Specify a filename for JSON runtime log output to be written to", typeof(bool));
             advancedGroup.AddArgument("psym-args", null, "Specify a concatenated list of additional PSym-specific arguments to pass, each starting with a colon").IsHidden = true;
             advancedGroup.AddArgument("jvm-args", null, "Specify a concatenated list of PSym-specific JVM arguments to pass, each starting with a colon").IsHidden = true;
             
@@ -213,9 +212,6 @@ namespace Plang.Options
                     break;
                 case "xml-trace": 
                     checkerConfiguration.IsXmlLogEnabled = true;
-                    break;
-                case "json-trace": 
-                    checkerConfiguration.IsJsonLogEnabled = true;
                     break;
                 case "explore":
                     checkerConfiguration.PerformFullExploration = true;
